refactor: 검색 결과 리스트 아이템 컴포넌트 분리 #6
